Fix limits enforce order when they're set at partition and other levels.
Previously it was taking the MIN, without respecting the order.

Also add a note to the resource_limits.html page to clarify the
exception for Max[Wall|Time] and/or [Max|Min]Nodes limits, where
the default is that the Partition is the king with regards of
precedence, unless the respective job's QOS flags
Partition[Min|Max|Time]Limit are set.

Bug 4681.