Do not release job resources twice when Cray NHC is stuck
When Cray NHC is stuck for a job for more than 300 seconds and a
reconfigure is issued, job resources will be released in order to
not keep all the job nodes allocated when there's no reason to.
If then NHC finishes, resources will be released again causing memory
under-allocated errors.

This patch avoids releasing resources more than one time.

Bug 4801