Only terminate processes if none are writing core.
Only one of a set of OpenMP processes will be marked as dumping
core, but terminating any of them will interrupt this and lead to
a partial core file being written.

Test all processes in the task, and only signal them if none are
writing out a core file. Otherwise wait and try again later.

Bug 4624.