Make UnavailableNodes value in job reason be correct for each job
1. Identifies nodes which are unavailable to a specific job,
adding a call to filter_by_node_owner() in select_nodes()
 where the node list is generated.
2. Removes the "unavail_node_str" argument to select_nodes()
as it is no longer useful. This string originally was originally
generated once at the start of the job scheduling logic for all jobs,
but since each job can have a different set of
 unavailable nodes (dedicated to user, group, etc.)
so the same string for all jobs can be misleading.

Bug 4932.