Set node->last_idle to 0 when in power_save state
In conjuction with previous commit (reconginizing nodes being powered up
out of band) set node's last_idle to 0 when the node is in a power_save
state. Additional meaning that the node isn't booted.

Partially reverts da722a891a. Checking for (last_idle > 0) when in
power_save state isn't necessary because if the node is already in
power_save state the node won't be resumed unless
(node_ptr->last_idle > (now - SuspendTime)). And with the previous
change, the node's last_idle time will be set when the node registers.