Add new parameters to GRES and add cons_tres plugin
The GRES data structures now support per-job, per-node, per-socket,
per-task parameters, plus cpus-per-gres and mem-per-gres. These
fields are not currently used, but the infrastructure is in place.
In addition there is a new cons_tres plugin, which is partly
functional, but needs a lot more work.
There are also some new tests, but slurm_unit and expect.