Fix regression from e346aa8a8eda5a0.
It turns out if we are running cache we need to call set_cluster_tres()
again since the pointers in the list have changed and don't have any
counts anymore.

I opted to just modify the existing locks and call it inside the lock
instead of having set_cluster_tres grab the locks.