Add create_mmap_buf() call to pack.c.
Given a filename, open() and mmap() it as the data field of a Buf.

Tag the Buf with 'mmaped' boolean to make sure free_buf knows
to munmap rather than xfree the data structure, and so we can
call fatal_abort if xfer_buf_data or grow_buf are called.