Fix srun to return highest signal of any task
srun would not return an exit code if a previous task exited before a
latter task exited with a signal.

If multiple tasks exit with a signal, srun returns the highest signal.

Partially reverts commit 04b449e1e63 -- the setting of local_global_rc
to NO_VAL as srun doesn't need to know whether it's been set or not
anymore. srun always sets the signal if a task exited with a signal.

Bug 5083