Remove select_g_select_jobinfo_get and HAVE_BG from job_mgr.c.
These calls to select_g_select_jobinfo_get() are a no-op
on non-select/bluegene systems, so remove them.

Since we're no longer modifying the job_desc through that call,
remove the debug3 statements around this as well.

The rest is removal of HAVE_BG blocks. Collapse the one else if
that was broken across an ifndef block as well here.