Remove AdminComment += syntax from 'scontrol update job'.
I do not see a use for this syntax, especially given that it appends
an extra comma in between the two halves. Only allow the full string
to change to put this in line with the Comment handling.

Remove special handling of an identical AdminComment as well,
since the end result is unchanged, and this avoids a potentially
expensive xstrcmp call.

Bug 5306.