cons_tres: remove bad max_cpus job limit
Original logic was testing against number of available CPUs on a
node, not the number actually being allocated. This can easily result
in a valid job not being possible to run. Note that setting the
max_cpus parameter requires the "scontrol update jobid=# numcpus=#-#"
command be explicitly executed as the option is not available at
job submit time, so it isn't something likely to happen on a
production system. Also note this bug exists in the cons_res plugin
and was just copied to cons_tres.