Improve escaping paths on user commands
When dealing with special characters like %A, %u, %s and so on and escaping it
on the command line, problems arises when one have directories with multiple
slashes in their names. This patch fixes this situation removing only one
slash on each pair of slashes just as normal escaping works i.e. in bash.

Bug 4859