X11 forwarding subsystem - add plumbing to permit a temporary XAUTHORITY file
Build out sufficient plumbing such that a temporary XAUTHORITY file
can be used that is local to the compute node, thus avoiding lock
contention on ~/.Xauthority on parallel filesystems.

This commit only includes the requisite plumbing to pass this around.

If this is not used, a null string results, and the XAUTHORITY env var
will not be forced into the user environment.

Add support and fix the modified API call in pam_slurm_adopt while here.

Bug 3647.