Print clustername when sacctmgr show runaway does
not find any jobs. The actual syntax for that sacctmgr command is

sacctmgr show runawayjobs [clustername]

but sometimes people mistakenly do

sacctmgr show runaway jobs

in which case it would look for runaway jobs on a cluster named "jobs"
and print the message "Runaway Jobs: No runaway jobs found" because the
clustername was wrong. This patch changes that message to say "Runaway
Jobs: No runaway jobs found on cluster jobs" so that people know they
made a mistake in syntax.