Use reliable communication with srun in _send_srun_resp_msg().
Do not use slurm_send_only_node_msg().

There is no way to tell if the srun has received the message before
the socket is shutdown if we do not wait to receive data. Use
slurm_send_recv_rc_msg_only_one() instead, and send back a response
from the other side.

We still need the older (and unreliable) behavior when talking to older
srun client commands, so make this change dependent on the protocol_version
field in the message.

Bug 5164.