Change debug messages() in _launch_handler().
After changes to slurm_send_only_node_msg(), this message is much
more likely to appear on systems with overloaded interconnects since
that connection handling code may end up retransmitting messages
that were actually received (but that the transmit side could not
verify were delivered successfully).

As the error() message stated, this isn't actually an error, and
the code will proceed happily past this point. So drop the debug
level, and remove the surrealist "this is not an error" part.

Bug 5164.