Rework logs involving _build_step_id().
Break JobId=... out from this, and only print the StepId=... bit.
This allows us to use %pJ in several additional places, and follows
the newer formatting used elsewhere of JobId=... StepId=... .

Due to how this code works, the end result is admittedly a bit uglier,
but the end result as printed to the logs is preferrable.