Clarify the two memory enforcement mechanisms
MemLimitEnforce and setting NoOverMemoryKill kills the job, or the job step
respectivelly if the memory is exceeded at some point based on JobAcctGather
plugin information, in a JobAcctGatherFreq frequency basis.

Both are incompatible with task/cgroup memory enforcement. The defaults have
been changed to not setting any mechanism, and info/fatal messages added
accordingly to give useful info to the admin. Moreover NoOverMemoryKill has
been deprecated in favor of OverMemoryKill and a fix on the detection of job
memory confinement has been included.

Bug 5479