Do not truncate MySQL database name at 33 characters.
MySQL permits up to 64-character database names, but Slurm was truncating
at 33-characters. If we exceed this limit, let the mysql_query fail and
give the admin a chance to sort it out, rather than truncating and then
failing to query against the un-truncated name later on.

While here correct the fatal() message.

Bug 5586.