Avoid slurmctld comp_job_cnt underflow error
This error occurs when one job is used to expand the allocation of
another job. The node record's "run_job_cnt" is decremented when
the dependent job's epilog completes and the job getting those
resources never has the "run_job_cnt" updated for it, which later
results in the "comp_job_cnt" underflow when it ends. This bug
was discovered in the course of select/cons_tres development,
but impacts all select plugins.