Change signature in backup.c to extern for (renamed) ping_controllers().
Both the primary and backup need to provide similar ping capabilities.
This implementation - with multiple threads launched in parallel - is
the more advanced one, so start reworking it to be able to use it in
the primary controller as well instead of _ping_backup_controller().

Add a boolean option to indicate whether this is running in the active
controller or the backup. This will control whether all other controllers
are pinged (needed by the primary) or only the higher priority controllers
(as done in the backups).

Bug 5256.