Change to a responding bool instead of response_time time_t value.
The handling for response_time was a bit odd, as care was needed to make
the timestamp values match up. Meaning they're not really time values,
just a complicated way of signaling whether the remote controller is
responding. So, simplify this by just indicating through a bool whether
a given controller is responding or not.

Bug 5256.