cons_tres enforce-binding fix
with some configurations test39.1 would fail. This was due to
not considering use of GRES that are non-local to available
CPUs, even without enforced binding. This was due to always
trying to enforce binding on the first pass. If that fails,
then consider use of non-local GRES. The problem in the logic
is it would never try non-local GRES if avail_res_array->avail_res_cnt
was zero. This patch sets its value to the node's CPU count,
which may not be ideal, but is a reasonable option.