add time limit to mpi step test
block distributions are failing with some configurations
ompi 4.0.x (head as of 10 sept 2018)
slurm 18.08
PMIx 3.0
The test with a block distribution is hanging. This will let the
remaining 3 MPI steps in test1.18 run to completion while we wait
for a fix from ompi.
Note that Alex and Danny are not seeing this failure, but they
might not be working off the latest ompi code.