front-end mulit-node deallocation
this is related to commit 845e134bf309b54a0f09255e02e38d6274e31613
when a step ends, flag that it is done on ALL nodes so that it's
memory can be released. Note that test1.19 would fail in front-end
mode when there was a default memory allocation. That test would
spawn 2 steps, but the first would not release its memory when
that step ended, so the second step would not start.