Docs - remove a few deprecated or non-recommended settings from configurator.html.
We do not recommend use of any AuthType aside from MUNGE. (And the links
here were already invalid.)

Similarly, CryptoType (and associated JobCredentialPrivateKey and
JobCredentialPublicCertificate options) are not generally used,
and should only be set after further consultation with the docs.

The only other option for CheckpointType besides checkpoint/none is
blcr which is deprecated, so remove that and JobCheckpointDir as well
while here.

Bug 5870.