Remove DEBUG_TIME blocks from slurm_cred.c.
I was going to rework these to remove the strcpy() calls, but
found that each of these have equivalent debug3() statements
just without the timestamps rendered as time strings.

Given that these are debug3(), and that you'd need to manually
inject a DEBUG_TIME definition somewhere to expose them, I don't
see the value of keeping them around. Especially as if you're digging
this far in to the source, it seems easier to troubleshoot off the raw
timestamp anyways IMNSHO.