Remove crypto/openssl plugin.
The only supported AuthType is auth/munge (the default). So running
crypto/munge (also the default) is the most natural solution, and minimizes
your exposure to security issues by using the same underlying API.

If you are using auth/none, then your system is insecure. Using
crypto/openssl to just avoid using crypto/munge does not buy you anything,
and you may as well use crypto/none and make it clear that your system
has no security in place. (The only caveat is that there is no crypto/none,
but that will be added in short order to fill in this gap.)

Bug 5916.