Don't consider partitions without enough nodes in reservation, main scheduler.
Main scheduler detects if a job asked for more than one partition and splits
one scheduling tasks for each partition. If the job asked for a reservation and
the intersection between the nodes in the reservation and the partition does
not suffice to satisfy the job, then skip scheduling this partition.

Otherwise this may cause useless scheduling effort and possibly overwriting
incorrectly the PD reason of a job.

bug 5895