Fix srun --prolog=none and --epilog=none to behave as documented
The srun man page states that --prolog=none or --epilog=none would
prevent the SrunProlog or SrunEpilog scripts defined in slurm.conf from
running. However, with SLURM_TEST_EXEC set in the environment, srun
would look for a script named "none" and run it if found. If not found,
then srun would print an error message, but the SrunProlog or
SrunEpilog would (correctly) not run.

* Prevent srun from running a script named "none" that is in the path.
* This also prevents the error message if it is not found.

Bug 5948.