Change to scontrol reporting of GRES binding
If a GRES is bound to all cores on a node rather than specific
cores on some sockets then "scontrol show node" will not show
the specific socket bounding. It will only report socket binding
for GRES that are bound to some subset of cores (and likely some
subset of sockets). This fixes the failure of test39.6 on our
knc node.