Rework x11 forwarding RPC fields.
Rename x11_target_host to x11_alloc_host to better indicate what the value
represents. Going forward, the x11_alloc_host and x11_alloc_port are the
hostname and TCP port number to connect to get the tunnel established.

The x11_target and x11_target_port fields indicate which X11 display to
connect to. If x11_target_port is zero, this indicates that x11_target
is a UNIX socket on x11_alloc_host. Otherwise, x11_target is the hostname
associated with the TCP port in x11_target_port for the DISPLAY.

Make careful changes to older protocol blocks to ensure the 17.11/18.08
slurmd processes can receive sufficient details from 19.05 slurmctld to
setup SSH-based forwarding.