handle change in gres count
if the gres count on a node with topology changes when the slurmctld
restarts then the gres data structures were left in an inconsistent
state. Namely the bitmaps would reflect the old size while the count
reflects the new size, which resulted in asserts. In addition, the
gres/mps data structure sizes need to match the gpu count on each
node. This new logic will synchronize mps data structures on gpu count
changes.