gres config work
This handles the case where some gres of a given Name have CPUs
configured for binding and others do not or have bad values.
It also fixes a bug in the conversion of that information to
a bitmap if the number of bits set is equal to the core count.