Remove slurmctld_comm_addr_t struct.
This is only used to define the slurmctld_comm_addr symbol in allocate.h.
Variable definitions do not belong in headers. Rather than relocate the
struct declaration, replace the struct which only holds a single port number
with a static variable serving the same purpose in allocate.c, and the
extern in this same location so srun.c can find it as well.