Fix initialization for lock arrays.
As pthread_rwlock_t is a struct, we cannot use this syntax
to initialize all elements of the array to the same value
as we would if these were any basic type.

Instead, move the initialization to a dedicated loop to ensure
all locks are initialized correctly.

This does not seem to have caused any issues on Linux platforms,
but macOS at least does notice this mistake.

Bug 6319.