Remove MemLimitEnforce parameter
This patch removes the MemLimitEnforce parameter from Slurm and moves the
condition to enable the feature to the JobAcctGatherParam=OverMemoryKill.
Setting OverMemoryKill will now enable the mechanism that was before enabled
with MemLimitEnforce=yes, so activation of killing jobs or steps due to
exceeding memory will be done in the same place. Also reconfiguring or
starting slurmd/ctld with incorrect configuration will cause a fatal error.

Bug 5479