Insure consistent GRES
If a gres type was identified in the node's GRES field, but that
type was not listed in the slurm.conf GresTypes parameter then
those GRES would appear in the output of "scontrol show node".
This misleading/wrong information would cause a bunch of tests
to fail. This patch only moves valid node GRES into the node
record and deletes the rest.