Add new function argument
This is needed so we can easily determine if we are running
in the slurmctld (rather than running in slurmd or in a test
program). When running with test7.3 and a configuration of
TopologyPlugin=topology/tree the test would generate a bunch of
"Invalid GRES configured on node ..."
errors and fail. This change means the GRES checking logic is
disabled for the test.