Use slurm.conf gres ids instead of gres.conf names to get a gres type name.
On a heterogeneous system the gres.conf coming from various nodes
might not be complete.  The slurm.conf has a complete list of possible
gres on the other hand.

If the GRES doesn't exist we will print an error and use the id instead
of the name, but this should never happen.

Bug 6366