Remove suspenders approach when dealing with people not enforcing associations.
What this code was used for was to try at the end of a job to see if
an association was there if there wasn't one there at the beginning of
the job.

From what we can tell the largest fallout here is a site wasn't
enforcing associations, then they do by scontrol reconfig and jobs
already running don't get an association id.  Since this was already
the case for any other job ran before hand this didn't seem like that
large of an issue.

What this does solve though is it allows you to release a job that was
held from a failed node since what was happening here was we got into
a state where if you ran

scontrol release 16862350_300
Job update not available right now, the DB index is being set, try again in a bit for job 16862350_300
slurm_suspend error: Job update not available right now, the DB index is being set, try again in a bit

This makes it so this state doesn't happen.

Bug 6340