Improve NVML library autodetection
Enable NVML lib and header to be found in system-default locations.
Add /usr/local/cuda/include and /usr/cuda/include to include path.
Remove NVML_LDFLAGS.
Remove the ability to specify a custom NVML path - this just
doesn't make sense with how the NVIDIA driver and CUDA are set up.
Get rid of searching for the lib in stubs/. Libraries in stubs/ are
only meant for cross-compiling and were never intended to be
dynamically linked against. Doing so produces a run-time warning and
may not even work.

Bug 5402