Change node selection with GrpNodes
If a GrpNodes limit is configurated in an association, partition QOS or
job QOS then favor use of nodes already allocated to that entity. This
will result in the configured node "Weight" being incremented by one for
nodes which are not prefered. Consider adjusting configured node "Weight"
values to achieve the desired node preferences.
bug 5303