Move plugin_type verification from auth plugins into slurm_auth.c.
Expose the plugin_type symbol from the plugins.

Ignore the plugin_version symbol, since on the unpack side we never do
anything with it. Just send a zero instead, and continue to ignore it
on the unpack side.

Note that for auth/munge, the 19.05 and the older protocol blocks are
identical, so collapse them back into one block.