Add index value to map the cred back in to a given auth plugin.
Use a new cred_wrapper_t struct designed to overlap the opaque cred
structures used within the individual plugins. Add the index value
to the start of each of these structures.

The index  is set when a given credential is unpacked, and used to lookup
the appropriate plugin for use with the remaining auth plugin
API calls.