Refactor auth plugins to remove the plugin_errno construct.
Use slurm_seterrno(), rather than maintaining a plugin-local and
credential local version of that errno.

Remove g_slurm_auth_errno() and g_slurm_auth_errstr(). Replace all
calling locations with %m format identifiers which should pick up
the errno in that thread, which is now directly set by the auth plugin.

The one place that this represents a slight regression in functionality
is that we loose the specific EMUNGE errno values. However, the places
that will set this in the auth/munge plugin are already printing it
through an error() with munge_ctx_strerror(). Replace those two locations
with ESLURM_AUTH_CRED_INVALID instead.