Add auth_index field to slurm_msg_t and set on after g_slurm_auth_unpack().
Identifies which authentication plugin (as identified by the index
rather than the plugin_id) was used to decode this specific
credential. Needed so that code paths responding to this message
can reply with the same authentication type to ensure the message
can be accepted by the remote (which, except for the slurmctld/slurmdbd,
will always be a single plugin type.