Remove SLURM_JOBID as an input variable to srun.
If running within salloc or a job step, both SLURM_JOB_ID and SLURM_JOBID
would be set to the same value, and SLURM_JOB_ID would overwrite SLURM_JOBID
as it is parsed second.

The only time this would cause a difference in behavior is if SLURM_JOBID
were set, and SLURM_JOB_ID were not. Which is not something that can be
caused by Slurm itself.