testsuite - remove test30.1 (rpmbuild).
IMO, the test as is currently designed is not representative of
an actual rpmbuild against a production released tarball, and thus
should not be relied on. Using --nodeps masks a lot of potential
errors with the spec file.

It has been disabled on Ubuntu systems for months, broken on master
for longer, and thus I am discarding it now.

Bug 4411, 6281.