Convert --mpi in srun to new form.
There is a mention in a3449c0b21 that it was pulled out of
slurm_opt_t previously because the Perl API needs it; I suspect the
path requiring it was removed with either BG/Q or Cray ALPS, and
this is no longer a strict requirement. At least there are no
references to mpi_type left there.