gres/mps fix for non-ordered gres/gpu
if the device files for gres/gpu are out of order or grouped
in an unordered fashion (e.g. "Name=gpu Files=/dev/nvidia[2,8,10]")
then split the gres/gpu records to one record per file and make
sure the gres/mps records are in an identical order. Required for
matching gres/gpu and gres/mps records (one GPU can be allocated
either as gres/gpu or as gres/mps, but not both, so we need to
be able to match records in slurmctld).