Fix issue with OR'ed job dependencies.
Job dependencies separated by "?" (OR'ed) should make the dependant job
be independent as soon as any of the dependencies are resolved to be
satisfied. Without this patch, if an invalid (non satisfiable)
dependency was resolved before a satisfiable one, then the dependant job
would never become independent, even after the satisfiable one got
eventually resolved.

Bug 5851.