Do not set CUDA_VISIBLE_DEVICES=NoDevFiles
In 0e149092 not setting the variable when job was not requesting any gres
was considered a bug. The cuda API will use all devices if the variable is
not set. If it is set to some unknown or empty value, it will use no devices.
This variable should be used only for testing purposes and ConstrainDevices=yes
in cgroup is recommended.

Bug 6412