Don't purge jobs if backfill is running.
The backfill scheduler keeps a local list of job pointers. Since the
backfill scheduler yields locks, it's possible for pending jobs to be
canceled and purged in these yield periods. The backfill scheduler then
has pointers to now invalid memory, and dereferencing those pointers is
undefined behavior and may result in a segfault.

This commit prevents purging jobs while the backfill scheduler is
running.

Bug 6621