[PATCH] Update GRES docs
Replace duplicated text with links, so text that should be the same
does not get out of sync.
Merge out-of-sync text together to get the best of both.
Fix error where line starting with 'nvml' was being omitted.
Minor grammatical and wording fixes.
Improve spacing of paragraphs.
Escape some missed `-` characters.
Remove some statements that are no longer true.
Sundry other minor changes.

An update in gres.conf.5 was not propagated to gres.shtml. Remove the
possibility of that happening again by simply referring the user to
the original doc via a link for more info.

Bug 4717