Correct bad test logic
When JobIDRaw started being published in the sacct -l output, the
result was to print the job ID twice, but part of the header being
looked for not at all. The result was for the test to pass, when
it really should not have. The bug was introduced in commit
9686a4b78ab, April 1, 2015.